+++ src/main/out_aud.h
@@ -4,8 +4,8 @@
 #include <stdint.h>
 
 void audio_mix(void *samples, size_t size);
-bool audio_init(float refreshra, float input_srate, float fps, bool fp);
-void audio_changeratefps(float refreshra, float input_srate, float fps);
+bool audio_init(double refreshra, float input_srate, double fps, bool fp);
+void audio_changeratefps(double refreshra, float input_srate, double fps);
 void audio_destroy();
 
 #endif
